requirements.txt
numpy==1.24.3
pandas==2.0.3
torch==2.0.1
torchvision==0.15.2
matplotlib==3.7.2
scikit-learn==1.3.0
aif360==0.5.0
tenseal==0.3.12
shap==0.42.1
streamlit==1.25.0
opencv-python==4.8.0.76
pillow==10.0.0
tqdm==4.65.0
requests==2.31.0
gdown==4.7.1
plotly==5.15.0 